Вынес большую часть  Game.cs в другие файлы и классы
